+++ fastimer/fastimer.py
@@ -9,6 +9,7 @@ from consolemenu.items import FunctionItem as Item
 from .show_fast import show_fast
 from .show_statistics import show_statistics
 from .start_fast import start_fast
+from .cancel_fast import cancel_fast
 from .stop_fast import stop_fast
 
 
@@ -21,6 +22,7 @@ def main() -> None:
 
     menu.append_item(Item("Start New Fast", start_fast, params))
     menu.append_item(Item("End Active Fast", stop_fast, params))
+    menu.append_item(Item("Cancel Active Fast", cancel_fast, params))
     menu.append_item(Item("Display Active Fast", show_fast, params))
     menu.append_item(Item("Display Fasting Statistics", show_statistics, params))